@@ -1,5 +1,3 @@
-import 'dart:ui';
-
 /// The "business rule" of the game. Depending on the count of neighbours,
 /// the [alive] changes.
 class CellState {
@@ -26,7 +24,7 @@ class Point {
   operator +(covariant Point other) => Point(x + other.x, y + other.y);
 
   @override
-  int get hashCode => hashValues(x, y);
+  int get hashCode => x ^ y;
 
   @override
   bool operator ==(Object other) {
@@ -47,7 +45,9 @@ class Grid {
     required this.yCount,
     List<Point> alivePoints = const [],
   }) {
-    alivePoints.forEach((point) => field[point] = true);
+    for (var point in alivePoints) {
+      field[point] = true;
+    }
   }
 
   final int xCount;
@@ -67,14 +67,14 @@ class Grid {
 
   /// List of the relative indices of the 8 cells around a cell.
   static List<Point> neighbourPoints = [
-    Point(-1, -1),
-    Point(0, -1),
-    Point(1, -1),
-    Point(-1, 0),
-    Point(1, 0),
-    Point(-1, 1),
-    Point(0, 1),
-    Point(1, 1),
+    const Point(-1, -1),
+    const Point(0, -1),
+    const Point(1, -1),
+    const Point(-1, 0),
+    const Point(1, 0),
+    const Point(-1, 1),
+    const Point(0, 1),
+    const Point(1, 1),
   ];
 }
 
@@ -99,12 +99,12 @@ class Game {
 class Patterns {
   static Grid beacon() {
     final alive = [
-      Point(1, 1),
-      Point(1, 2),
-      Point(2, 1),
-      Point(4, 3),
-      Point(3, 4),
-      Point(4, 4),
+      const Point(1, 1),
+      const Point(1, 2),
+      const Point(2, 1),
+      const Point(4, 3),
+      const Point(3, 4),
+      const Point(4, 4),
     ];
     return Grid(xCount: 6, yCount: 6, alivePoints: alive);
   }
@@ -112,57 +112,57 @@ class Patterns {
   static Grid pulsar() {
     final alive = [
       // top left
-      Point(4, 2),
-      Point(5, 2),
-      Point(6, 2),
-      Point(2, 4),
-      Point(2, 5),
-      Point(2, 6),
-      Point(4, 7),
-      Point(5, 7),
-      Point(6, 7),
-      Point(7, 4),
-      Point(7, 5),
-      Point(7, 6),
+      const Point(4, 2),
+      const Point(5, 2),
+      const Point(6, 2),
+      const Point(2, 4),
+      const Point(2, 5),
+      const Point(2, 6),
+      const Point(4, 7),
+      const Point(5, 7),
+      const Point(6, 7),
+      const Point(7, 4),
+      const Point(7, 5),
+      const Point(7, 6),
       // top right
-      Point(10, 2),
-      Point(11, 2),
-      Point(12, 2),
-      Point(9, 4),
-      Point(9, 5),
-      Point(9, 6),
-      Point(10, 7),
-      Point(11, 7),
-      Point(12, 7),
-      Point(14, 4),
-      Point(14, 5),
-      Point(14, 6),
+      const Point(10, 2),
+      const Point(11, 2),
+      const Point(12, 2),
+      const Point(9, 4),
+      const Point(9, 5),
+      const Point(9, 6),
+      const Point(10, 7),
+      const Point(11, 7),
+      const Point(12, 7),
+      const Point(14, 4),
+      const Point(14, 5),
+      const Point(14, 6),
       // bottom left
-      Point(4, 9),
-      Point(5, 9),
-      Point(6, 9),
-      Point(2, 10),
-      Point(2, 11),
-      Point(2, 12),
-      Point(4, 14),
-      Point(5, 14),
-      Point(6, 14),
-      Point(7, 10),
-      Point(7, 11),
-      Point(7, 12),
+      const Point(4, 9),
+      const Point(5, 9),
+      const Point(6, 9),
+      const Point(2, 10),
+      const Point(2, 11),
+      const Point(2, 12),
+      const Point(4, 14),
+      const Point(5, 14),
+      const Point(6, 14),
+      const Point(7, 10),
+      const Point(7, 11),
+      const Point(7, 12),
       // bottom right
-      Point(10, 9),
-      Point(11, 9),
-      Point(12, 9),
-      Point(9, 10),
-      Point(9, 11),
-      Point(9, 12),
-      Point(10, 14),
-      Point(11, 14),
-      Point(12, 14),
-      Point(14, 10),
-      Point(14, 11),
-      Point(14, 12),
+      const Point(10, 9),
+      const Point(11, 9),
+      const Point(12, 9),
+      const Point(9, 10),
+      const Point(9, 11),
+      const Point(9, 12),
+      const Point(10, 14),
+      const Point(11, 14),
+      const Point(12, 14),
+      const Point(14, 10),
+      const Point(14, 11),
+      const Point(14, 12),
     ];
     return Grid(xCount: 17, yCount: 17, alivePoints: alive);
   }
@@ -170,70 +170,70 @@ class Patterns {
   static Grid pentaDecathlon() {
     final alive = [
       // top
-      Point(5, 3),
-      Point(4, 4),
-      Point(5, 4),
-      Point(6, 4),
-      Point(3, 5),
-      Point(4, 5),
-      Point(5, 5),
-      Point(6, 5),
-      Point(7, 5),
+      const Point(5, 3),
+      const Point(4, 4),
+      const Point(5, 4),
+      const Point(6, 4),
+      const Point(3, 5),
+      const Point(4, 5),
+      const Point(5, 5),
+      const Point(6, 5),
+      const Point(7, 5),
       // bottom
-      Point(3, 12),
-      Point(4, 12),
-      Point(5, 12),
-      Point(6, 12),
-      Point(7, 12),
-      Point(4, 13),
-      Point(5, 13),
-      Point(6, 13),
-      Point(5, 14),
+      const Point(3, 12),
+      const Point(4, 12),
+      const Point(5, 12),
+      const Point(6, 12),
+      const Point(7, 12),
+      const Point(4, 13),
+      const Point(5, 13),
+      const Point(6, 13),
+      const Point(5, 14),
     ];
     return Grid(xCount: 11, yCount: 18, alivePoints: alive);
   }
 
   static Grid gosperGliderGun() {
     final alive = [
-      Point(1, 5),
-      Point(2, 5),
-      Point(1, 6),
-      Point(2, 6),
+      const Point(1, 5),
+      const Point(2, 5),
+      const Point(1, 6),
+      const Point(2, 6),
       //
-      Point(11, 5),
-      Point(11, 6),
-      Point(11, 7),
-      Point(12, 4),
-      Point(12, 8),
-      Point(13, 3),
-      Point(13, 9),
-      Point(14, 3),
-      Point(14, 9),
-      Point(15, 6),
-      Point(16, 4),
-      Point(16, 8),
-      Point(17, 5),
-      Point(17, 6),
-      Point(17, 7),
-      Point(18, 6),
+      const Point(11, 5),
+      const Point(11, 6),
+      const Point(11, 7),
+      const Point(12, 4),
+      const Point(12, 8),
+      const Point(13, 3),
+      const Point(13, 9),
+      const Point(14, 3),
+      const Point(14, 9),
+      const Point(15, 6),
+      const Point(16, 4),
+      const Point(16, 8),
+      const Point(17, 5),
+      const Point(17, 6),
+      const Point(17, 7),
+      const Point(18, 6),
       //
-      Point(21, 3),
-      Point(21, 4),
-      Point(21, 5),
-      Point(22, 3),
-      Point(22, 4),
-      Point(22, 5),
-      Point(23, 2),
-      Point(23, 6),
-      Point(25, 1),
-      Point(25, 2),
-      Point(25, 6),
-      Point(25, 7),
+      const Point(21, 3),
+      const Point(21, 4),
+      const Point(21, 5),
+      const Point(22, 3),
+      const Point(22, 4),
+      const Point(22, 5),
+      const Point(23, 2),
+      const Point(23, 6),
+      const Point(25, 1),
+      const Point(25, 2),
+      const Point(25, 6),
+      const Point(25, 7),
       //
-      Point(35, 3),
-      Point(35, 4),
-      Point(36, 3),
-      Point(36, 4),
+      const Point(35, 3),
+      const Point(35, 4),
+      const Point(36, 3),
+      const Point(36, 4),
     ];
     return Grid(xCount: 38, yCount: 12, alivePoints: alive);
   }
